CODE RUB: Code Cleanup
